+++ src/io/ft_putuint_fd.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putuint_fd.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: iwordes <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: iwordes <iwordes@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/16 10:53:43 by iwordes           #+#    #+#             */
-/*   Updated: 2016/11/16 10:53:44 by iwordes          ###   ########.fr       */
+/*   Updated: 2016/12/14 19:10:09 by iwordes          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "ft_io.h"
+#include <libft.h>
 
-void	ft_putuint_fd(unsigned long long integer, int fd)
+void	ft_putuint_fd(uintmax_t integer, int fd)
 {
 	ft_putuint_base_fd(integer, 10, fd);
 }
